Add tests for `with_subsequence`, `is_subsequence`, and `intersect` (#31)
